feat: 실시간 공시 알림 기능 구현 및 안정성 강화
- DART API를 이용한 주기적 공시 확인 기능 구현 (APScheduler)
- DB 기반(system_config) 중복 알림 방지 로직 추가
- 사용자 정의 예외(DartApiError) 도입 등 API 통신 오류 처리 강화
- 관련 기능 문서화(docs/features) 및 PLAN.md 현행화

+++ src/api/models/daily_price.py
@@ -1,15 +1,15 @@
-from sqlalchemy import Column, Integer, String, Float, Date, DateTime, func
-from src.api.models.base import Base
+from sqlalchemy import Column, String, DateTime, Float, BigInteger, Date, func
+from src.common.db_connector import Base
 
 class DailyPrice(Base):
     __tablename__ = 'daily_prices'
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     symbol = Column(String(20), index=True, nullable=False)
     date = Column(Date, index=True, nullable=False)
     open = Column(Float, nullable=False)
     high = Column(Float, nullable=False)
     low = Column(Float, nullable=False)
     close = Column(Float, nullable=False)
-    volume = Column(Integer, nullable=False)
+    volume = Column(BigInteger, nullable=False)
     created_at = Column(DateTime, default=func.now(), nullable=False)
     updated_at = Column(DateTime, default=func.now(), onupdate=func.now(), nullable=False)
