+++ src/api/models/prediction_history.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, DateTime, func
-from src.api.models.base import Base
+from sqlalchemy import Column, BigInteger, String, Float, DateTime, ForeignKey, func
+from src.common.db_connector import Base
 from datetime import datetime
 
 class PredictionHistory(Base):
     __tablename__ = 'prediction_history'
-    id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, nullable=False)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    user_id = Column(BigInteger, nullable=False)
     symbol = Column(String(20), nullable=False)
     prediction = Column(String(20), nullable=False)
     created_at = Column(DateTime, default=func.now(), nullable=False)
